pyproject.toml
[project]
name = "met-office-check"
version = "0.1.0"
description = "Check the Met Office data in the Planetary Computer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "obstore>=0.9.2",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pystac>=1.14.3",
    "rustac>=0.9.8",
    "stactools-met-office-deterministic>=0.4.2",
    "tqdm>=4.67.3",
    "typer>=0.24.1",
]

[project.scripts]
met-office-check = "met_office_check:app"

[build-system]
requires = ["uv_build>=0.11.0,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
